# 03-tmcra-web-console/public/openapi.json
{
  "components": {
    "schemas": {
...
      "ConsistencyContract": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "enum": [
              "eventual",
              "read_your_writes"
            ],
            "title": "Mode",
            "type": "string"
          },
          "recall_wait_for_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recall Wait For Job Id"
          },
          "visible_after_job_id": {
            "title": "Visible After Job Id",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "visible_after_job_id"
        ],
        "title": "ConsistencyContract",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "title": "Details"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "retry_after_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Seconds"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EvidenceRouteView": {
        "additionalProperties": false,
        "properties": {
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "requested": {
            "title": "Requested",
            "type": "string"
          },
          "selected": {
            "enum": [
              "raw",
              "compiled"
            ],
            "title": "Selected",
            "type": "string"
          }
        },
        "required": [
          "requested",
          "selected",
          "reasons"
        ],
        "title": "EvidenceRouteView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestRequest": {
        "additionalProperties": false,
        "properties": {
          "consistency": {
            "default": "eventual",
            "enum": [
              "eventual",
              "read_your_writes"
            ],
            "title": "Consistency",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MemoryMessage"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "session_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          },
          "slow_policy": {
            "default": "auto",
            "enum": [
              "auto",
              "deferred",
              "force"
            ],
            "title": "Slow Policy",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "messages"
        ],
        "title": "IngestRequest",
        "type": "object"
      },
...
      "JobView": {
        "additionalProperties": false,
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "consistency_contract": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsistencyContract"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "heartbeat_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeat At"
          },
          "idempotent_replay": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotent Replay"
          },
          "idempotent_retry": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotent Retry"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "job_type": {
            "title": "Job Type",
            "type": "string"
          },
          "lease_expires_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lease Expires At"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "resume_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Mode"
          },
          "scope_name": {
            "title": "Scope Name",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "number"
          }
        },
        "required": [
          "job_id",
          "tenant_id",
          "scope_name",
          "job_type",
          "status",
          "attempts",
          "created_at",
          "updated_at",
          "status_url"
        ],
        "title": "JobView",
        "type": "object"
      },
...
      "MemoryMessage": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 200000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "message_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Message Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "role": {
            "enum": [
              "user",
              "assistant",
              "system",
              "tool"
            ],
            "title": "Role",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "role",
          "content",
          "timestamp"
        ],
        "title": "MemoryMessage",
        "type": "object"
      },
      "PromptEvidenceView": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "content_character_count": {
            "title": "Content Character Count",
            "type": "integer"
          },
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "format": {
            "enum": [
              "text/plain",
              "application/json"
            ],
            "title": "Format",
            "type": "string"
          },
          "memory_context_block_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Context Block Count"
          },
          "mode": {
            "enum": [
              "raw_hierarchical",
              "compiled_evidence_packet"
            ],
            "title": "Mode",
            "type": "string"
          },
          "neighbor_block_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neighbor Block Count"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "source_block_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Block Count"
          },
          "source_text_verbatim": {
            "title": "Source Text Verbatim",
            "type": "boolean"
          },
          "trust_boundary": {
            "title": "Trust Boundary",
            "type": "string"
          },
          "window_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Count"
          }
        },
        "required": [
          "schema_version",
          "format",
          "mode",
          "content",
          "content_sha256",
          "content_character_count",
          "source_text_verbatim",
          "trust_boundary"
        ],
        "title": "PromptEvidenceView",
        "type": "object"
      },
...
      "RecallRequest": {
        "additionalProperties": false,
        "properties": {
          "debug": {
            "default": false,
            "title": "Debug",
            "type": "boolean"
          },
          "evidence_mode": {
            "default": "auto",
            "enum": [
              "raw",
              "auto",
              "compiled"
            ],
            "title": "Evidence Mode",
            "type": "string"
          },
          "max_windows": {
            "const": 8,
            "default": 8,
            "title": "Max Windows",
            "type": "integer"
          },
          "query": {
            "maxLength": 100000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "query_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Time"
          },
          "recall_profile": {
            "default": "quality",
            "enum": [
              "quality",
              "interactive"
            ],
            "title": "Recall Profile",
            "type": "string"
          },
          "response_projection": {
            "default": "full",
            "enum": [
              "full",
              "prompt_only"
            ],
            "title": "Response Projection",
            "type": "string"
          },
          "wait_for_job_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wait For Job Id"
          }
        },
        "required": [
          "query"
        ],
        "title": "RecallRequest",
        "type": "object"
      },
      "RecallResponse": {
        "additionalProperties": false,
        "properties": {
          "debug": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debug"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "evidence_route": {
            "$ref": "#/components/schemas/EvidenceRouteView"
          },
          "index_job_id": {
            "title": "Index Job Id",
            "type": "string"
          },
          "prompt_evidence": {
            "$ref": "#/components/schemas/PromptEvidenceView"
          },
          "query_id": {
            "title": "Query Id",
            "type": "string"
          },
          "scope_name": {
            "title": "Scope Name",
            "type": "string"
          }
        },
        "required": [
          "query_id",
          "scope_name",
          "index_job_id",
          "evidence_route",
          "evidence",
          "prompt_evidence"
        ],
        "title": "RecallResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "TMCRAApiKey": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Tenant-isolated long-term memory ingestion and prompt-ready recall. TMCRA does not generate the final assistant answer.",
    "title": "TMCRA Memory API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/scopes/{scope_name}/ingest": {
      "post": {
        "operationId": "ingestMemory",
        "parameters": [
          {
            "in": "path",
            "name": "scope_name",
            "required": true,
            "schema": {
              "title": "Scope Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "description": "Trusted subject attribution for server-side control-plane proxies. Requires a tokens:manage API key.",
            "in": "header",
            "name": "X-TMCRA-On-Behalf-Of-Subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Trusted subject attribution for server-side control-plane proxies. Requires a tokens:manage API key.",
              "title": "X-Tmcra-On-Behalf-Of-Subject"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "TMCRAApiKey": []
          }
        ],
        "summary": "Ingest",
        "tags": [
          "memory"
        ]
      }
    },
...
    "/v1/scopes/{scope_name}/recall": {
      "post": {
        "operationId": "recallMemory",
        "parameters": [
          {
            "in": "path",
            "name": "scope_name",
            "required": true,
            "schema": {
              "title": "Scope Name",
              "type": "string"
            }
          },
          {
            "description": "Trusted subject attribution for server-side control-plane proxies. Requires a tokens:manage API key.",
            "in": "header",
            "name": "X-TMCRA-On-Behalf-Of-Subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Trusted subject attribution for server-side control-plane proxies. Requires a tokens:manage API key.",
              "title": "X-Tmcra-On-Behalf-Of-Subject"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecallResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Tenant quota, rate, or recall queue limit was reached.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Recall capacity is temporarily unavailable.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "TMCRAApiKey": []
          }
        ],
        "summary": "Recall",
        "tags": [
          "memory"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "TMCRA Memory API",
      "url": "https://api.tmcra.com"
    }
  ],
...
}